Fix 152: Fix input bug in beamdp/beamdp_gui for option 11.
Previously, beamdp_gui only allowed the user to input no.
of particles for which to output parameters.  This did
not reflect inputs actually required by beamdp: the range
of particles for which to list parameters.  Confusion was
exacerbated by the command line description of this input
from beamdp, itself, which also indicated only the no.
of particles for which to output data.  This has been
fixed as well.  Also, fixed a serious bug resulting in
erroneous data output for all options when IAEA phase space
data was analyzed.