Fix issue when GASP>0 and density correction file used.
Calculated density corrections based on an input value of GASP
(gas pressure) are disabled when a density correction file is used.
Moreover, if both a density correction file and a GASP>0 are specified,
the pegsles implementation crashes if GASP is not unity.  In the fix,
GASP is reset to 0 (i.e. corrections using GASP turned off) if a
density correction file is used.